익명 댓글 기능 삭제 시 비밀번호 체크 기능 추가

+++ board-backend/src/main/java/com/example/board/controller/CommentGuestController.java
@@ -2,6 +2,8 @@ package com.example.board.controller;
 
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -10,8 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.example.board.dto.CommentDto;
 import com.example.board.dto.CommentGuestDto;
+import com.example.board.dto.PasswordDto;
+import com.example.board.entity.CommentGuest;
 import com.example.board.service.CommentGuestService;
 
 import lombok.RequiredArgsConstructor;
@@ -35,12 +38,33 @@ public class CommentGuestController {
 		return commentGuestService.getComments(boardId);
 	}
 	
-    // 댓글 삭제
+    // 비밀번호 체크 후 댓글 삭제
     @DeleteMapping("/{commentId}")
-    public void deleteComment(
+    public ResponseEntity<?> deleteComment(
     		@PathVariable("boardId") Long boardId, 
-    		@PathVariable("commentId") Long commentId) {
+    		@PathVariable("commentId") Long commentId, 
+    		@RequestBody PasswordDto dto) {
+    	Boolean check = commentGuestService.checkPassword(commentId, dto);
+    	
+    	if (!check) {
+    		return ResponseEntity.status(HttpStatus.FORBIDDEN).body("비밀번호가 불일치 합니다.");
+    	}
+    	
         commentGuestService.deleteComment(commentId);
+        
+        return ResponseEntity.ok("삭제 완료");
+    }
+    
+    // 댓글 비밀번호 검증
+    @DeleteMapping("/{commentId}/check")
+    public ResponseEntity<Boolean> checkPassword(
+    		@PathVariable("boardId") Long boardId,
+    		@PathVariable("commentId") Long commentId,
+    		@RequestBody PasswordDto dto) {
+    	
+    	Boolean chk = commentGuestService.checkPassword(commentId, dto);
+    	
+    	return ResponseEntity.ok(chk);
     }
 
 }
